+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.6.0 (2020-10-??)
 
+### 🚀 New features
+
+* [#34](https://github.com/antoinezanardi/werewolves-assistant-web/issues/34) - Users can review a game with rating and comment.
+
 ### 🌟 Enhancements
 
 * [#30](https://github.com/antoinezanardi/werewolves-assistant-web/issues/30) - `.eslintrc.js` file contains now all existing rules. Rules were adapted for each type of files.
@@ -14,12 +18,12 @@
 
 * `express-rate-limit` installed with version `5.1.3`.
 * `xss` installed with version `1.0.8`.
-* `@sentry/node` updated to version `5.24.2`.
+* `@sentry/node` updated to version `5.25.0`.
 * `apidoc` updated to version `5.24.2`.
 * `eslint` updated to version `7.10.0`.
 * `migrate-mongo` updated to version `8.1.2`.
 * `mocha` updated to version `8.1.3`.
-* `mongoose` updated to version `5.10.7`.
+* `mongoose` updated to version `5.10.9`.
 * `validator` updated to version `13.1.17`.
 
 ---